apps/api/otel_setup.py
from __future__ import annotations

import logging
import os

from fastapi import FastAPI

logger = logging.getLogger(__name__)


def setup_otel_fastapi(app: FastAPI) -> None:
    """
    Optional OpenTelemetry OTLP HTTP export (works with self-hosted Phoenix).
    Set OTEL_EXPORTER_OTLP_ENDPOINT, e.g. http://localhost:6006/v1/traces
    """
    endpoint = os.getenv("OTEL_EXPORTER_OTLP_ENDPOINT")
    if not endpoint:
        return
    try:
        from opentelemetry import trace
        from opentelemetry.exporter.otlp.proto.http.trace_exporter import OTLPSpanExporter
        from opentelemetry.instrumentation.fastapi import FastAPIInstrumentor
        from opentelemetry.sdk.resources import Resource
        from opentelemetry.sdk.trace import TracerProvider
        from opentelemetry.sdk.trace.export import BatchSpanProcessor

        resource = Resource.create({"service.name": os.getenv("OTEL_SERVICE_NAME", "saas-copilot-api")})
        provider = TracerProvider(resource=resource)
        exporter = OTLPSpanExporter(endpoint=endpoint)
        provider.add_span_processor(BatchSpanProcessor(exporter))
        trace.set_tracer_provider(provider)
        FastAPIInstrumentor.instrument_app(app)
        logger.info("OpenTelemetry OTLP enabled: %s", endpoint)
    except Exception as exc:
        logger.warning("OpenTelemetry setup skipped: %s", exc)
