apps/api/requirements.txt
fastapi
uvicorn
pydantic
httpx
qdrant-client
langgraph
duckdb
langfuse
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp-proto-http
